[jovian] extend `extraData` to include minBaseFee in EIP1559Params (#580)
<!--
Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.

Contributors guide:
https://github.com/alloy-rs/core/blob/main/CONTRIBUTING.md

The contributors guide includes instructions for running rustfmt and
building the
documentation.
-->

<!-- ** Please select "Allow edits from maintainers" in the PR Options
** -->

From: https://github.com/wlawt/op-alloy/pull/1

## Motivation

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.
-->

This PR adds the necessary helper functions to encode/decode the
[configurable minimum base
fee](https://github.com/ethereum-optimism/design-docs/blob/main/protocol/minimum-base-fee.md)
feature to op-alloy.


Specifically, this extends the `extraData` section in the block header
to include an extra uint64 for this addition. The default value would be
zero, as if there was no minimum base fee to begin with if the
configurable minimum base fee wasn't enabled.

By introducing a minimum base fee, it reduces the chances of
experiencing priority fee auctions which occur when the block is almost
at the gas limit, as well as cases where batcher sequencer throttling is
happening where it intentionally processes fewer transactions to avoid a
backlog of batches.

## Solution

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->

By setting a minimum base fee, we reduce the time and the number of
doublings needed. See the table on the time saved on [OP Mainnet in
minutes](https://github.com/ethereum-optimism/design-docs/blob/main/protocol/minimum-base-fee.md#minimum-base-fee-in-block-header).

## PR Checklist

- [x] Added Tests
- [x] Added Documentation
- [ ] Breaking changes